+++ activesupport/test/notifications_test.rb
@@ -462,21 +462,6 @@ module Notifications
       assert_equal Hash[payload: :bar], event.payload
     end
 
-    def test_event_is_parent_based_on_children
-      time = Process.clock_gettime(Process::CLOCK_MONOTONIC)
-
-      parent    = event(:foo, Process.clock_gettime(Process::CLOCK_MONOTONIC), Process.clock_gettime(Process::CLOCK_MONOTONIC) + 100, random_id, {})
-      child     = event(:foo, time, time + 10, random_id, {})
-      not_child = event(:foo, time, time + 100, random_id, {})
-
-      parent.children << child
-
-      assert parent.parent_of?(child)
-      assert_not child.parent_of?(parent)
-      assert_not parent.parent_of?(not_child)
-      assert_not not_child.parent_of?(parent)
-    end
-
     def test_subscribe_raises_error_on_non_supported_arguments
       notifier = ActiveSupport::Notifications::Fanout.new